Room: added derived TransferDefaultRoom
This room class should be instanced only once and called to processed at the end of the chain in order to receive all superfluous UMMU transfers. I.e. all UMMU transfers to rooms that are already full.
Since it should be the last in the chain, it will also define the "vessel full" marker and therefore logically represent the vessel's "single" crew-compartment.